Support an unbounded zip header
Previously we'd search the first 64kb for the zip in order to determine
how to decode the file, but it turns out that zips orient themselves by
the end of the file so we can take advantage that any zip parser worth
its salt will ignore (and indeed, expose) the number of bytes in the
header.

Doing this allowed the code to be shorter without a performance hit.
Additionally an unknown bug was solved with this method. And there is
one less dependency (not that `twoway` was very big to begin with)