Melt with tabs instead of spaces
While I am a fan of spaces, this commit switches the output of the
melted output to tabs for a two reasons:

The game outputs saves with tabs so we should try and follow that format
where possible.

And most importantly, tabs represent a good tradeoff between readability
and space efficiency. Before using the official write API, a single
space was used to increase the depth of the document. This occassionally
made it hard for me to visually distinguish separate parts of the
document. My first thought was to use two spaces instead of one but this
increased the melted output of one of my EU4 1.31 saves by 15MB (20%).
This is a bit unacceptable so tabs is used as a compromise.