Update for CK3 1.5
CK3 1.5 changed the floating point encoding. This required new plumbing
in the underlying parser as previously it assumed that floating point
format would only vary between games not between game patches. So this
commit detects the floating point format by looking at the first couple
binary tokens to determine.

Having the deserialization change based on the detected floating point
is still a todo, and that's fine for now as this crate is mostly used
for melting.

The other change in this commit is that the API has been expanded to
accept customizing token stringification.