Tweak melter for 1.4 full compatibility
Before tackling the backwards incompatibilities with ck3 1.5, I first
needed a strong regression test for ck3 1.4 saves. The regression test
involves comparing the melted output with a known good melted output.
The known good melted output was generated by comparing a debug_mode
cloud save to an autosave.

To get to the known good output, some tweaks needed to be made to the
current melter:

- Tone down the zealousness of writing all integers in
  `ai_strategies_index` as numbers (only write keys)
- `damage_last_tick` is a known number
- Fixes the small precision issue with negative numbers